feat: user can register and admin view registration
